# docs/contracts/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Merge Game API",
    "description": "Unity \uBA38\uC9C0 \uAC8C\uC784 \uD074\uB77C\uC774\uC5B8\uD2B8\uB97C \uC704\uD55C \uC11C\uBC84 \uAD8C\uC704\uD615 HTTP API\uC785\uB2C8\uB2E4.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/players/me": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "\uAC80\uC99D\uB41C JWT\uC758 sub \uD074\uB808\uC784\uC5D0 \uD574\uB2F9\uD558\uB294 \uD604\uC7AC \uD50C\uB808\uC774\uC5B4 \uD504\uB85C\uD544\uC744 \uBC18\uD658\uD569\uB2C8\uB2E4.",
        "operationId": "GetCurrentPlayer",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentPlayerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentPlayerResponse": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "string",
            "description": "JWT\uC640 \uC77C\uCE58\uD558\uB294 \uD50C\uB808\uC774\uC5B4 \uC2DD\uBCC4\uC790\uC785\uB2C8\uB2E4.",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "description": "\uD604\uC7AC \uAC8C\uC784 \uD45C\uC2DC \uC774\uB984\uC785\uB2C8\uB2E4.",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "description": "\uACC4\uC815 \uC0DD\uC131 UTC \uC2DC\uAC01\uC785\uB2C8\uB2E4.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "\uC778\uC99D\uB41C \uD604\uC7AC \uD50C\uB808\uC774\uC5B4\uC758 \uACF5\uAC1C \uD504\uB85C\uD544 \uC751\uB2F5\uC785\uB2C8\uB2E4."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "POST /api/v1/auth/guest\uC5D0\uC11C \uBC1C\uAE09\uBC1B\uC740 accessToken\uC744 \uC785\uB825\uD569\uB2C8\uB2E4.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AdminApiKey": {
        "type": "apiKey",
        "description": "\uC6B4\uC601 \uBE44\uBC00 \uC800\uC7A5\uC18C\uC5D0\uC11C \uC8FC\uC785\uD55C \uAD00\uB9AC\uC790 API \uD0A4\uC785\uB2C8\uB2E4. Unity \uD074\uB77C\uC774\uC5B8\uD2B8\uC5D0\uB294 \uD3EC\uD568\uD558\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4.",
        "name": "X-Admin-Key",
        "in": "header"
      }
    }
  }
}